# backend/openapi/swagger.yaml
components:
  schemas:
...
    github_com_uptaris_uptaris_backend_internal_inventory.IncidentRow:
      properties:
        createdAt:
          type: string
        description:
          maxLength: 10000
          type: string
        id:
          example: "0"
          type: string
        monitorId:
          example: "0"
          type: string
        monitorName:
          type: string
        resolvedAt:
          type: string
        serverId:
          example: "0"
          type: string
        severity:
          enum:
          - low
          - medium
          - high
          - critical
          type: string
        startedAt:
          type: string
        status:
          enum:
          - open
          - acknowledged
          - resolved
          type: string
        title:
          maxLength: 200
          type: string
        updatedAt:
          type: string
      required:
      - severity
      - startedAt
      - status
      - title
      type: object
    github_com_uptaris_uptaris_backend_internal_models.Incident:
      properties:
        createdAt:
          type: string
        description:
          maxLength: 10000
          type: string
        id:
          example: "0"
          type: string
        monitorId:
          example: "0"
          type: string
        resolvedAt:
          type: string
        severity:
          enum:
          - low
          - medium
          - high
          - critical
          type: string
        startedAt:
          type: string
        status:
          enum:
          - open
          - acknowledged
          - resolved
          type: string
        title:
          maxLength: 200
          type: string
        updatedAt:
          type: string
      required:
      - severity
      - startedAt
      - status
      - title
      type: object
...
    github_com_uptaris_uptaris_backend_internal_request.IncidentInput:
      properties:
        description:
          example: Primary customer API
          maxLength: 10000
          type: string
        resolvedAt:
          example: "2026-09-26T11:00:00Z"
          type: string
        severity:
          enum:
          - low
          - medium
          - high
          - critical
          example: low
          type: string
        startedAt:
          example: "2026-09-26T10:00:00Z"
          type: string
        status:
          enum:
          - open
          - acknowledged
          - resolved
          example: open
          type: string
        title:
          example: API maintenance
          maxLength: 200
          type: string
      required:
      - severity
      - title
      type: object
    github_com_uptaris_uptaris_backend_internal_request.IncidentPatchInput:
      properties:
        description:
          example: Primary customer API
          maxLength: 10000
          type: string
        resolvedAt:
          example: "2026-09-26T11:00:00Z"
          type: string
        severity:
          enum:
          - low
          - medium
          - high
          - critical
          example: low
          type: string
        startedAt:
          example: "2026-09-26T10:00:00Z"
          type: string
        status:
          enum:
          - open
          - acknowledged
          - resolved
          example: open
          type: string
        title:
          example: API maintenance
          maxLength: 200
          type: string
      type: object
...
    github_com_uptaris_uptaris_backend_internal_response.ErrorResponse:
      properties:
        error:
          properties:
            code:
              type: string
            message:
              type: string
          required:
          - code
          - message
          type: object
      required:
      - error
      type: object
    github_com_uptaris_uptaris_backend_internal_response.IncidentListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_models.Incident'
          type: array
          uniqueItems: false
        links:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.Links'
        pagination:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.Pagination'
      type: object
    github_com_uptaris_uptaris_backend_internal_response.IncidentOverviewResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_inventory.IncidentRow'
          type: array
          uniqueItems: false
        links:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.Links'
        pagination:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.Pagination'
      type: object
    github_com_uptaris_uptaris_backend_internal_response.IncidentResponse:
      properties:
        _links:
          additionalProperties:
            type: string
          type: object
        data:
          $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_models.Incident'
      type: object
    github_com_uptaris_uptaris_backend_internal_response.Links:
      properties:
        next:
          type: string
        previous:
          type: string
        self:
          type: string
      type: object
...
    github_com_uptaris_uptaris_backend_internal_response.Pagination:
      properties:
        page:
          example: 1
          type: integer
        pageSize:
          example: 20
          type: integer
        total:
          example: 5
          type: integer
        totalPages:
          example: 1
          type: integer
      type: object
...
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  description: REST API for Uptaris infrastructure monitoring.
  title: Uptaris API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /incidents:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        schema:
          default: 1
          minimum: 1
          type: integer
      - description: Page size
        in: query
        name: pageSize
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
      - description: Status
        in: query
        name: status
        schema:
          enum:
          - open
          - acknowledged
          - resolved
          type: string
      - description: Severity
        in: query
        name: severity
        schema:
          enum:
          - low
          - medium
          - high
          - critical
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.IncidentOverviewResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_query: invalid query parameter'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: List accessible incidents across servers
      tags:
      - incidents
...
  /servers/{serverId}/monitors/{monitorId}/incidents:
    get:
      parameters:
      - description: Server ID
        in: path
        name: serverId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Monitor ID
        in: path
        name: monitorId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Page number
        in: query
        name: page
        schema:
          default: 1
          minimum: 1
          type: integer
      - description: Results per page
        in: query
        name: pageSize
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
      - description: Severity
        in: query
        name: severity
        schema:
          enum:
          - low
          - medium
          - high
          - critical
          type: string
      - description: Status
        in: query
        name: status
        schema:
          enum:
          - open
          - acknowledged
          - resolved
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.IncidentListResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_id or invalid_query: invalid ID or query parameter'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'not_found: resource not found'
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: List monitor incidents
      tags:
      - incidents
    post:
      parameters:
      - description: Server ID
        in: path
        name: serverId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Monitor ID
        in: path
        name: monitorId
        required: true
        schema:
          minimum: 1
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_request.IncidentInput'
        description: Incident fields
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_models.Incident'
          description: Created
          headers:
            Location:
              description: Created resource URI
              schema:
                type: string
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_id or invalid_json: invalid ID or JSON body'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'forbidden or origin_not_allowed: role or origin not permitted'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'not_found: resource not found'
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "415":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'unsupported_media_type: non-JSON request body'
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'validation_failed: invalid request field'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: Create monitor incident
      tags:
      - incidents
  /servers/{serverId}/monitors/{monitorId}/incidents/{incidentId}:
    delete:
      parameters:
      - description: Server ID
        in: path
        name: serverId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Monitor ID
        in: path
        name: monitorId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Incident ID
        in: path
        name: incidentId
        required: true
        schema:
          minimum: 1
          type: integer
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_id: invalid resource ID'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'forbidden or origin_not_allowed: role or origin not permitted'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'not_found: resource not found'
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "415":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'unsupported_media_type: non-JSON request body'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: Delete incident
      tags:
      - incidents
    get:
      parameters:
      - description: Server ID
        in: path
        name: serverId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Monitor ID
        in: path
        name: monitorId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Incident ID
        in: path
        name: incidentId
        required: true
        schema:
          minimum: 1
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.IncidentResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_id: invalid resource ID'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'not_found: resource not found'
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: Get incident
      tags:
      - incidents
    patch:
      parameters:
      - description: Server ID
        in: path
        name: serverId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Monitor ID
        in: path
        name: monitorId
        required: true
        schema:
          minimum: 1
          type: integer
      - description: Incident ID
        in: path
        name: incidentId
        required: true
        schema:
          minimum: 1
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_request.IncidentPatchInput'
        description: Fields to update
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_models.Incident'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'invalid_id or invalid_json: invalid ID or JSON body'
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: authentication_required, invalid_token, revoked_token, or stale_token
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'forbidden or origin_not_allowed: role or origin not permitted'
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'not_found: resource not found'
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'body_too_large: request body exceeds 64 KiB'
        "415":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'unsupported_media_type: non-JSON request body'
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'validation_failed: invalid request field'
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'internal_error: unexpected server error'
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_uptaris_uptaris_backend_internal_response.ErrorResponse'
          description: 'service_unavailable: database operation failed'
      security:
      - bearerauth: []
      summary: Update incident
      tags:
      - incidents
...
servers:
- url: /api/v1